Kl-159/refactor: seperate input to component
- input을 새로운 컴포넌트로 분리해서 rerender 방지
