+++ src/pages/SubmitPage/components/InfoSubmit/InfoSubmit.jsx
@@ -1,19 +1,16 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Input, Rate } from 'antd'
-import { useShallow } from 'zustand/react/shallow'
 
 import useKyQuery from '../../../../hooks/useKyQuery'
-import useFormStore from '../../../../stores/useFormStore'
 import theme from '../../../../styles/theme'
+import NameInput from './NameInput'
+import DescriptionInput from './DescriptionInput'
+import RatingInput from './RatingInput'
 import NumberInputForm from './NumberInputForm'
 
 export default function InfoSubmitForm() {
-  const [name, description, rating] = useFormStore(
-    useShallow((state) => [state.name, state.description, state.rating])
-  )
-  const setFormContents = useFormStore((state) => state.setFormContents)
   const { data: currencies, isLoading } = useKyQuery('currencies')
+  console.log(currencies)
 
   return (
     <>
@@ -23,30 +20,11 @@ export default function InfoSubmitForm() {
       </h2>
       <Wrapper>
         <h3>상품명</h3>
-        <Input
-          showCount
-          maxLength={50}
-          defaultValue={name}
-          placeholder="상품명을 입력해주세요"
-          onBlur={(e) => setFormContents({ name: e.target.value.trim() })}
-          style={{ fontSize: theme.size.textSM }}
-          size="large"
-        />
+        <NameInput />
       </Wrapper>
       <Wrapper>
         <h3>후기</h3>
-        <Input.TextArea
-          showCount
-          maxLength={1000}
-          defaultValue={description}
-          placeholder="리뷰를 입력해 주세요"
-          onBlur={(e) =>
-            setFormContents({ description: e.target.value.trim() })
-          }
-          size="large"
-          autoSize={{ minRows: 3 }}
-          style={{ fontSize: theme.size.textSM }}
-        />
+        <DescriptionInput />
       </Wrapper>
       <Wrapper>
         <h3>구매가격</h3>
@@ -54,19 +32,14 @@ export default function InfoSubmitForm() {
       </Wrapper>
       <Wrapper>
         <h3>별점</h3>
-        <Rate
-          allowHalf
-          defaultValue={rating}
-          style={{ color: '#FFD700' }}
-          onChange={(num) => setFormContents({ rating: num })}
-        />
+        <RatingInput />
       </Wrapper>
     </>
   )
 }
 
 const Wrapper = styled.div`
-  width: 600px;
+  width: 37.5rem;
   position: relative;
   margin-bottom: 1.875rem;
